Aula 2

+++ app/Http/Controllers/UsersController.php
@@ -5,13 +5,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Routing\Route;
+use App\User;
 
 class UsersController extends Controller
 {
 
     public function index()
     {
-        $users = \App\User::all();
+        $users = User::all();
 
         return view('user.index')->with(compact('users'));
     }
@@ -23,7 +24,7 @@ class UsersController extends Controller
 
     public function store(Request $form)
     {
-        $user = new \App\User();
+        $user = new User();
         //Atribui valores do formulário ao ORM User
         $user->name = $form->input('name');
         $user->email = $form->input('email');
@@ -31,36 +32,44 @@ class UsersController extends Controller
         //Salva o dados no db.
         $user->save();
 
-        return redirect()->route('users.index');
+        return redirect()->route('users.index')->with('message','Item criado com sucesso!');
     }
 
-    public function edit(Request $form)
+    public function edit($id = null)
     {
-        $user = \App\User::find($form->input('id'));
-
-        return view('user.edit')->with(compact('user'));
+        $user = User::find($id);
+        if (!is_null($user)):
+            return view('user.edit')->with(compact('user', 'id'));
+        else:
+            return redirect()->route('users.index')->withErrors(['Usuário não encontrado!']);
+        endif;
     }
 
-    public function update(Request $form)
+    public function update(Request $form, $id)
     {
         //Assing ORM  from bd from id = ?
-        $user = \App\User::find($form->input('id'));
+        $user = User::find($id);
         //Set new values in ORM
         $user->name = $form->input('name');
         $user->email = $form->input('email');
         //Save ORM in database
         $user->update();
 
-        return redirect()->route('users.index');
+
+        return redirect()->route('users.index')->with('message','Item atualizado com sucesso!');
     }
 
-    public function delete(Request $form)
+    public function delete($id = null)
     {
         //Assing ORM  from bd from id = ?
-        $user = \App\User::find($form->input('id'));
+        $user = User::find($id);
         //Delete database file
-        $user->delete();
+        if(!empty($user)):
+            $user->delete();
+            return redirect()->route('users.index')->with('message','Item deletado!');
+        else:
+            return redirect()->route('users.index')->withErrors(['Não foi possível deletar!']);
+        endif;
 
-        return redirect()->route('users.index');
     }
 }
